requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.0.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.20.0
setuptools==70.3.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
